balance fixes 2
-massive balance changes(among enemies, level cfgs, units)
-fixed some bugs(raycast with t.right, not be able to hit enemy on some maps)
-wawes amount now can be seen

+++ Assets/_Project/Scripts/InGame/GUI/ShowDamageText.cs
@@ -5,7 +5,7 @@ public class ShowDamageText : MonoBehaviour
     public DynamicTextData TextData;
     [SerializeField] private float _offsetY;
 
-    public void ShowDamage(float dmg)
+    public void ShowDamage(int dmg)
     {
         Vector2 newVec = new(transform.position.x, transform.position.y + _offsetY);
         DynamicTextManager.CreateText2D(newVec, dmg.ToString(), TextData);
